requirements.txt
discord.py==2.0.1
jishaku==2.5.1
asyncpg==0.26.0